checkHoliday.js: Moved to src from tests
